Fix issue with speed cost being incorrectly calculated
The problem was that the speed tracking service was not properly handling the case where it was deactivated twice in a row. This was causing it to erroneously consider extra time as having speed on. This fix adds proper handling for this case which seems to have fixed the overall issue with speed cost.