Bugfix for broadcaster
The problem turned out to be in this component which batches messages destined for the frontend together. Under certain circumstances, it would collect messages but would never send them to the frontend at the end of the round. When this happened to a reference price message, the displayed reference price (and by extension the PBD) would be incorrect.

I fixed this by ensuring that batches are always sent properly. I suspect this bug was also causing some other weird frontend behaviors, it would have affected more than just reference price changes.